requirements.txt
fastapi
uvicorn
pandas
python-dotenv
langchain
langchain-core
langgraph
pinecone-client
sentence-transformers
openai
snowflake-connector-python
matplotlib
serpapi
